refactor: rename supabase utility middleware.ts -> proxy.ts
Matches Next.js 16 naming convention where the root proxy.ts now imports
from '@/lib/supabase/proxy' instead of '@/lib/supabase/middleware'.
Updates index.ts barrel export accordingly.
