+++ src/lib/supabase/index.ts
@@ -1,6 +1,6 @@
 export { createClient as createBrowserClient, getClient } from './client';
 export { createClient as createServerClient } from './server';
-export { updateSession } from './middleware';
+export { updateSession } from './proxy';
 export { mapSupabaseReviewToReview, mapSupabaseReviewsToArray } from './mappers';
 export type { SupabaseReview } from './mappers';
 export type { Database } from './database.types';
